% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalize.R
\name{along_track}
\alias{along_track}
\title{Colocalize along a cruise track.  Takes a cruise name and colocalizes
  the cruise track with the specified variable(s). THIS IS THE MAIN
  COLOCALIZATION function for a cruise.}
\usage{
along_track(cruise, targetTables, targetVars, depth1, depth2,
  temporalTolerance, latTolerance, lonTolerance, depthTolerance)
}
\arguments{
\item{cruise}{String of the name of cruise. You can see the list of cruises
using \code{cruises()}.}

\item{depth1}{LOWER limit for depth along the cruise track (lat/lon
trajectory) the user is interested in.}

\item{depth2}{UPPER limit for depth along the cruise track (lat/lon
trajectory) the user is interested in.}
}
\value{
A dataset colocalized with the cruise.
}
\description{
Colocalize along a cruise track.  Takes a cruise name and colocalizes
  the cruise track with the specified variable(s). THIS IS THE MAIN
  COLOCALIZATION function for a cruise.
}
