% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscSubroutine.R
\name{any.overlap}
\alias{any.overlap}
\title{Helper to identify overlap between one range (table.range) and the reference
(source.range).}
\usage{
\method{any}{overlap}(table.range, source.range)
}
\arguments{
\item{table.range}{Lat/lon/time range of target table.}

\item{source.range}{Lat/lon/time range of source.}
}
\value{
\code{TRUE} if any overlap exists in lat/lon/time.
}
\description{
Helper to identify overlap between one range (table.range) and the reference
(source.range).
}
