% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscSubroutine.R
\name{any_overlap}
\alias{any_overlap}
\title{Helper to identify overlap between one range (table_range) and the reference
(source_range).}
\usage{
any_overlap(table_range, source_range)
}
\arguments{
\item{table_range}{Lat/lon/time range of target table.}

\item{source_range}{Lat/lon/time range of source.}
}
\value{
\code{TRUE} if any overlap exists in lat/lon/time.
}
\description{
Helper to identify overlap between one range (table_range) and the reference
(source_range).
}
