% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalize.R
\name{atomic_match}
\alias{atomic_match}
\title{Colocalizes the source variable (from source table) with a single target
long variable (from target table).  The tolerance parameters set the
matching boundaries between the source and target data sets.  Returns a
dataframe containing the source variable joined with the target variable.}
\usage{
atomic_match(spName, sourceTable, sourceVar, targetTable, targetVar, dt1,
  dt2, lat1, lat2, lon1, lon2, depth1, depth2, temporalTolerance,
  latTolerance, lonTolerance, depthTolerance)
}
\arguments{
\item{sourceTable}{table name of the source data set.}

\item{dt1}{start date or datetime.}

\item{dt2}{end date or datetime.}

\item{lat1}{start latitude [degree N].}

\item{lat2}{end latitude [degree N].}

\item{lon1}{start longitude [degree E].}

\item{lon2}{end longitude [degree E].}

\item{depth1}{start depth [m].}

\item{depth2}{end depth [m].}

\item{latTolerance}{float list of spatial tolerance values in meridional
direction [deg] between pairs of source and target data sets. If only one
value is given, that would be applied to all target data sets.}

\item{lonTolerance}{float list of spatial tolerance values in zonal
direction [deg] between pairs of source and target data sets. If only one
value is given, that would be applied to all target data sets.}

\item{depthTolerance}{float list of spatial tolerance values in vertical
direction [m] between pairs of source and target data sets. If only one
value is given, that would be applied to all target data sets.}

\item{spname}{stored procedure name that executes the matching logic.}

\item{sourceVariable}{the source variable. The target variables are matched
(colocalized) with this variable.}

\item{targetTables}{table names of the target data sets to be matched with
the source data.}

\item{targetVariables}{variable names to be matched with the source
variable.}

\item{timeTolerance}{float list of temporal tolerance values between pairs
of source and target datasets. The size and order of values in this list
should match those of targetTables. If only a single integer value is
given, that would be applied to all target datasets. This parameter is in
day units except when the target variable represents monthly climatology
data in which case it is in month units. Notice fractional values are not
supported in the current version.}
}
\description{
Colocalizes the source variable (from source table) with a single target
long variable (from target table).  The tolerance parameters set the
matching boundaries between the source and target data sets.  Returns a
dataframe containing the source variable joined with the target variable.
}
