% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_vartype.R
\docType{data}
\name{catalog_vartype}
\alias{catalog_vartype}
\title{Variable type from the catalog}
\format{A tibble object of variable type: 
\describe{
  \item{catalog_vartype}{Variable type catalog.}
}}
\usage{
data(catalog_vartype)
}
\description{
Table representing tableName, variableName, and its data type.
}
\details{
Data frame catalog_vartype, contain tableName, variableName, and its data type.
}
\examples{
# data(catalog_vartype)
}
\keyword{datasets}
