% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorization.R
\name{check_authorization}
\alias{check_authorization}
\title{Check query execution}
\usage{
check_authorization(query_out)
}
\arguments{
\item{query_out}{output after a query}
}
\description{
Check query execution
}
