% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{check_with_catalog}
\alias{check_with_catalog}
\title{Helper to check variable name or table name with the catalog.}
\usage{
check_with_catalog(thing, type = c("varName", "tableName"), apiKey)
}
\description{
Helper to check variable name or table name with the catalog.
}
