% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{columns}
\alias{columns}
\title{Returns the list of columns of a data set.}
\usage{
columns(tableName)
}
\arguments{
\item{tableName}{Table name.}
}
\description{
Returns the list of columns of a data set.
}
\examples{
\dontrun{
columns('tblAMT13_Chisholm')
}
}
