% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect2cmap}
\alias{connect2cmap}
\title{Connect to the CMAP database}
\usage{
connect2cmap(Driver = "libtdsodbc.so")
}
\arguments{
\item{Driver}{name of the database driver to be used to connect to the database}
}
\value{
connection object for the database
}
\description{
User needs to provide CMAP databse login credentials once to connect to the database. Thereafter, call "connect2cmap" with driver inofrmation to connect to the database.
}
\examples{
\dontrun{
con <- connect2cmap(Driver = "libtdsodbc.so")
#
## Input: Table name; variable name, space time range information
table.name <- "tblsst_AVHRR_OI_NRT" # table name
sel.var <- "sst" # choose variable
range.var <- list() # Range variable [lat,lon,time]
range.var$lat <- c(10, 70)
range.var$lon <- c(-180, -80)
range.var$time <- c("2016-04-30", "2016-04-30")
#
## Subset selection:
tbl.subset <- getTableData(con, table.name, sel.var, range.var)
head(tbl.subset)
#
dbDisconnect(con)
}
}
