% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{cruise_trajectory}
\alias{cruise_trajectory}
\title{Returns a dataframe containing the cruise trajectory (Identical function in python)}
\usage{
cruise_trajectory(cruisename)
}
\description{
Returns a dataframe containing the cruise trajectory (Identical function in python)
}
