% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cruise_info.R, R/interface.R
\name{cruise_variables}
\alias{cruise_variables}
\title{Returns a dataframe containing all registered variables (at Simons CMAP) during a cruise.}
\usage{
cruise_variables(cruiseName)

cruise_variables(cruiseName)
}
\arguments{
\item{cruiseName}{name of cruise.}
}
\description{
Returns a dataframe containing all registered variables (at Simons CMAP) during a cruise.

Returns a dataframe containing all registered variables (at Simons CMAP) during a cruise.
}
\examples{
\donttest{
library(cmap4r)
#
## Variable attribute:
cruiseVar <- cruise_variables('SCOPE_Falkor1')
#
}
}
