% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{cruises}
\alias{cruises}
\title{Returns a dataframe containing a list of all of the hosted cruise names.}
\usage{
cruises()
}
\description{
Returns a dataframe containing a list of all of the hosted cruise names.
}
