% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{depth_profile}
\alias{depth_profile}
\title{Retrieve depth profile data from  CMAP using space time range parameters}
\usage{
depth_profile(table, variable, dt1, dt2, lat1, lat2, lon1, lon2, depth1,
  depth2)
}
\arguments{
\item{table_name}{table name in the database}

\item{sel_var}{select a variable in the table}

\item{range_var}{range of time, latitude, longitude, depth}
}
\value{
subset of a table as data frame
}
\description{
Get subset of a table as data frame
}
\examples{
\dontrun{
library(cmap4r)
#
## Input: Table name; variable name, space time range information
table = "tblArgoMerge_REP"
variable = "argo_merge_chl_adj"
lat1 = 20
lat2 = 24
lon1 = -170
lon2 = -159
dt1 = "2016-04-30"
dt2 = "2016-04-30"
depth1 = 0
depth2 = 1500
depthprofile(table, variable, dt1, dt2, lat1, lat2, lon1, lon2, depth1, depth2)
#
}
}
