% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{exec_manualquery}
\alias{exec_manualquery}
\title{Retrieve data using manual query.}
\usage{
exec_manualquery(manual_query)
}
\arguments{
\item{manual_query}{own query of the user.}
}
\value{
subset of a table as data frame
}
\description{
Get subset of a table as data frame
}
\examples{
\donttest{
library(cmap4r)
#
manual_query = "SELECT [time], lat, lon, depth, Fe FROM tblPisces_NRT
  WHERE
[time] BETWEEN '2017-06-03' AND '2017-06-03' AND
lat BETWEEN 10 AND 55 AND
lon BETWEEN -180 AND 100 AND
depth BETWEEN 0 AND 0.5
ORDER BY [time], lat, lon, depth"
#
## Subset selection:
tbl.subset <- exec_manualquery(manual_query)
head(tbl.subset)
#
}
}
