% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_cmap.R
\name{get_aggtable}
\alias{get_aggtable}
\title{Retrieve aggregated subset of table on CMAP}
\usage{
get_aggtable(con, table.name, sel.var, range.var, agg.var,
  order.var = NULL)
}
\arguments{
\item{con}{connection object to the database}

\item{table.name}{table name in the database}

\item{sel.var}{select a variable in the table}

\item{range.var}{range of time, latitude, longitude, depth}

\item{agg.var}{aggregate variable}

\item{order.var}{retrieve data with order by variable}
}
\value{
aggregated subset of a table
}
\description{
Get aggregated subset of a table
}
\examples{
\dontrun{
library(cmap4r)
con <- connect_cmap(Driver = "libtdsodbc.so")
#
## Input: Table name; variable name, space time range information
table.name <- "tblsst_AVHRR_OI_NRT" # table name
sel.var <- "sst" # choose variable
range.var <- list() # Range variable [lat,lon,time]
range.var$lat <- c(25, 30)
range.var$lon <- c(-160, -155)
range.var$time <- c("2016-03-29", "2016-05-29")
#
# Aggregate
agg.var <- "time" # Specify aggregate variable
tbl.subset <- get_aggtable(con, table.name, sel.var, range.var, agg.var)
head(tbl.subset)
#
dbDisconnect(con)
}
}
