% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R, R/table_info.R
\name{get_catalog}
\alias{get_catalog}
\title{Returns catalog of the CMAP data base.}
\usage{
get_catalog()

get_catalog()
}
\value{
table of catalog

table of catalog
}
\description{
Returns catalog of the CMAP data base.

Returns catalog of the CMAP data base.
}
\examples{
\dontrun{
library(cmap4r)
#
#
## Variable attribute:
cmap.catalog <- get_catalog()
cmap.catalog
#
}
\donttest{
library(cmap4r)
#
## Variable attribute:
cmap.catalog <- get_catalog()
cmap.catalog
#
}
}
