% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_info.R
\name{get_columns}
\alias{get_columns}
\title{Returns the list of columns of a data set.}
\usage{
get_columns(tableName)
}
\arguments{
\item{tableName}{table name.}
}
\value{
columns in a table
}
\description{
Returns the list of columns of a data set.
}
\examples{
\donttest{
library(cmap4r)
#
## Input: Table name;
table_name <- "tblAMT13_Chisholm" # table name
#
## Subset selection:
tbl.columns <- get_columns(table_name)
tbl.columns
#
}
}
