% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cruise_info.R
\name{get_cruise_trajectory}
\alias{get_cruise_trajectory}
\title{Returns a dataframe containing the cruise trajectory (Identical function in python)}
\usage{
get_cruise_trajectory(cruisename)
}
\arguments{
\item{cruisename}{name of cruise.}
}
\description{
Returns a dataframe containing the cruise trajectory (Identical function in python)
}
\examples{
\donttest{

#
## Variable attribute:
cruiseTrajectory  <- get_cruise_trajectory('KM1513')
#
}
}
