% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_depthprofile}
\alias{get_depthprofile}
\title{Retrieve depth profile data from  CMAP using space time range parameters}
\usage{
get_depthprofile(table_name, sel_var, range_var)
}
\arguments{
\item{table_name}{table name in the database}

\item{sel_var}{select a variable in the table}

\item{range_var}{range of time, latitude, longitude, depth}
}
\value{
subset of a table as data frame
}
\description{
Get subset of a table as data frame
}
\examples{
\dontrun{
library(cmap4r)
#
## Input: Table name; variable name, space time range information
table_name <- "tblArgoMerge_REP" # table name
sel_var <- "argo_merge_chl_adj" # choose variable
range_var <- list() # Range variable [lat,lon,time]
range_var$lat <- c(20, 24)
range_var$lon <- c(-170, -150)
range_var$time <- c("2016-04-30", "2016-04-30")
range_var$depth <- c(0, 1500)
#
## Subset selection:
tbl.subset <- get_depthprofile(table_name, sel_var, range_var)
head(tbl.subset)
#
}
}
