% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_info.R
\name{get_head}
\alias{get_head}
\title{Returns top  n records  from a table on the Simons CMAP.}
\usage{
get_head(tableName, nrows = 5)
}
\arguments{
\item{tableName}{table name from the Simons CMAP database. Use "get_catalog()" to retrieve list of tables on the database.}

\item{nrows}{number of rows to retrieve.}
}
\value{
return table  as dataframe
}
\description{
Returns top  n records  from a table on the Simons CMAP.
}
\examples{
\donttest{

## Input: Table name;
tableName <- "tblArgoMerge_REP" # table name
#
## Top n rows:
tbl.subset <- get_head(tableName, nrows=10)
tbl.subset
}
}
