% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{get_metadata}
\alias{get_metadata}
\title{Returns a dataframe containing the associated metadata. The inputs can be
strings (if only one table, and variable is passed) or a list of
string literals.}
\usage{
get_metadata(tables, variables)
}
\arguments{
\item{tables}{Vector of table names.}

\item{variables}{Vector of variable names in the tables.}
}
\description{
Returns a dataframe containing the associated metadata. The inputs can be
strings (if only one table, and variable is passed) or a list of
string literals.
}
\examples{
\dontrun{
table <- c("tblArgoMerge_REP") # table name
variable <- c("argo_merge_chl_adj") # variable name
get_metadata(table, variable)
}
}
