% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_references}
\alias{get_references}
\title{Returns a dataframe containing references associated with a data set.
datasetID ID of dataset.}
\usage{
get_references(datasetID)
}
\description{
Returns a dataframe containing references associated with a data set.
datasetID ID of dataset.
}
