% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_cmap.R
\name{get_table}
\alias{get_table}
\title{Retrieve subset of data from a table on CMAP}
\usage{
get_table(con, table.name, sel.var, range.var, order.var = NULL)
}
\arguments{
\item{con}{connection object to the database}

\item{table.name}{table name in the database}

\item{sel.var}{select a variable in the table}

\item{range.var}{range of time, latitude, longitude, depth}

\item{order.var}{retrieve data with order by variable}
}
\value{
subset of a table as data frame
}
\description{
Get subset of a table as data frame
}
\examples{
\dontrun{
library(cmap4r)
con <- connect_cmap(Driver = "libtdsodbc.so")
#
## Input: Table name; variable name, space time range information
table.name <- "tblsst_AVHRR_OI_NRT" # table name
sel.var <- "sst" # choose variable
range.var <- list() # Range variable [lat,lon,time]
range.var$lat <- c(10, 70)
range.var$lon <- c(-180, -80)
range.var$time <- c("2016-04-30", "2016-04-30")
#
## Subset selection:
tbl.subset <- get_table(con, table.name, sel.var, range.var)
head(tbl.subset)
#
dbDisconnect(con)
}
}
