% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_timeseries}
\alias{get_timeseries}
\title{Retrieve timeseries data from  CMAP using space time range parameters}
\usage{
get_timeseries(table_name, sel_var, range_var, interval = NULL)
}
\arguments{
\item{table_name}{table name in the database}

\item{sel_var}{select a variable in the table}

\item{range_var}{range of time, latitude, longitude, depth}

\item{interval}{choose time series data aggregation at [weekly, montly, quaterly, annual, none]}
}
\value{
subset of a table as data frame
}
\description{
Get subset of a table as data frame
}
\examples{
\dontrun{
library(cmap4r)
#
## Input: Table name; variable name, space time range information
table_name <- "tblHOT_Bottle" # table name
sel_var <- "SiO4_bottle_hot" # choose variable
range_var <- list() # Range variable [lat,lon,time]
range_var$lat <- c(22, 23)
range_var$lon <- c(-159, -157)
range_var$time <- c("1988-12-01", "2016-10-15")

# Subset selection:
tbl.subset <- get_timeseries(table_name, sel_var, range_var)
head(tbl.subset)
#
}
}
