% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{get_unit}
\alias{get_unit}
\title{Returns the unit for a given variable.}
\usage{
get_unit(tableName, varName)
}
\arguments{
\item{tableName}{Table name.}

\item{varName}{Variable name.}
}
\description{
Returns the unit for a given variable.
}
\examples{
\dontrun{
tableName <- "tblArgoMerge_REP" # table name
varName <- "argo_merge_chl_adj" # variable name
get_unit(tableName, varName)
}
}
