% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{get_var_catalog}
\alias{get_var_catalog}
\title{Returns a single-row dataframe from catalog (udfCatalog) containing all of
the variable's info at catalog.}
\usage{
get_var_catalog(tableName, varName)
}
\arguments{
\item{tableName}{Table name.}

\item{varName}{Variable name.}
}
\value{
Single-row data frame.
}
\description{
Returns a single-row dataframe from catalog (udfCatalog) containing all of
the variable's info at catalog.
}
\examples{
\dontrun{
tableName <- "tblArgoMerge_REP" # table name
varName <- "argo_merge_chl_adj" # variable name
get_var_catalog(tableName, varName)
}
}
