% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_info.R
\name{get_var_coverage}
\alias{get_var_coverage}
\title{Returns a single-row dataframe from the database  catalog containing the
variable's spatial and temporal coverage.}
\usage{
get_var_coverage(tableName, varName)
}
\arguments{
\item{tableName}{table name.}

\item{varName}{variable name.}
}
\value{
spatio-temporal range information of a table variable as dataframe
}
\description{
Returns a single-row dataframe from the database  catalog containing the
variable's spatial and temporal coverage.
}
\examples{
\donttest{
#
## Input: Table name; variable name
tableName <- "tblArgoMerge_REP" # table name
varName <- "argo_merge_chl_adj" # variable name
#
## Variable attribute:
varCoverage <- get_var_coverage(tableName, varName)
varCoverage
#
}
}
