% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{get_var_resolution}
\alias{get_var_resolution}
\title{Returns a single-row dataframe from catalog (udfCatalog) containing the
variable's spatial and temporal resolutions.}
\usage{
get_var_resolution(tableName, varName)
}
\arguments{
\item{tableName}{Table name.}

\item{varName}{Variable name.}
}
\description{
Returns a single-row dataframe from catalog (udfCatalog) containing the
variable's spatial and temporal resolutions.
}
\examples{
\dontrun{
tableName <- "tblArgoMerge_REP" # table name
varName <- "argo_merge_chl_adj" # variable name
get_var_resolution(tableName, varName)
}
}
