% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_info.R
\name{get_var_stat}
\alias{get_var_stat}
\title{Returns a single-row dataframe from catalog (udfCatalog) containing the variable's summary statistics.}
\usage{
get_var_stat(tableName, varName)
}
\arguments{
\item{tableName}{table name.}

\item{varName}{variable name.}
}
\description{
Returns a single-row dataframe from catalog (udfCatalog) containing the variable's summary statistics.
}
\examples{
\donttest{
library(cmap4r)
#
## Input: Table name; variable name
table_name <- "tblArgoMerge_REP" # table name
var_name <- "argo_merge_chl_adj" # variable name
#
## Variable attribute:
varStats <- get_var_stat(table_name, var_name)
varStats
#
}
}
