% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_info.R
\name{get_var_unit}
\alias{get_var_unit}
\title{Returns unit of the variable.}
\usage{
get_var_unit(tableName, varName)
}
\arguments{
\item{tableName}{table name in the database}

\item{varName}{name of variable}
}
\value{
subset of a table as data frame
}
\description{
Returns unit of the variable.
}
\examples{
\donttest{
library(cmap4r)
#
## Input: Table name; variable name
table_name <- "tblArgoMerge_REP" # table name
var_name <- "argo_merge_chl_adj" # variable name
#
## Variable attribute:
unitName <- get_var_unit(table_name, var_name)
unitName
#
}
}
