% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_cmap.R
\name{getAggregatedTableData}
\alias{getAggregatedTableData}
\title{Retrieve aggregated subset of table on CMAP}
\usage{
getAggregatedTableData(con, table.name, sel.var, range.var, agg.var,
  order.var = NULL)
}
\arguments{
\item{con}{connection object to the database}

\item{table.name}{table name in the database}

\item{sel.var}{select a variable in the table}

\item{range.var}{range of time, latitude, longitude, depth}

\item{agg.var}{aggregate variable}

\item{order.var}{retrieve data with order by variable}
}
\value{
aggregated subset of a table
}
\description{
Get aggregated subset of a table
}
\examples{
\dontrun{
con <- dbConnect(odbc::odbc(), DSN="CMAP-MSSQL",UID="ArmLab",PWD="ArmLab2018")
#
## Input: Table name; variable name, space time range information
table.name = 'tblsst_AVHRR_OI_NRT'    # table name
sel.var = 'sst'                         # choose variable
range.var <- list()                     # Range variable [lat,lon,time]
range.var$lat <- c(25,30)
range.var$lon <- c(-160,-155)
range.var$time <- c('2016-03-29', '2016-05-29')
#
# Aggregate
agg.var <- 'time'                       # Specify aggregate variable
tbl.subset <- getAggregatedTableData(con, table.name, sel.var, range.var, agg.var)
head(tbl.subset)
#
dbDisconnect(con)
}
}
