% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableFeature.R
\name{getColClass}
\alias{getColClass}
\title{Get class of columns in the table}
\usage{
getColClass(con, table.name)
}
\arguments{
\item{con}{connection object to the database}

\item{table.name}{table name in the database}
}
\value{
column name with its class
}
\description{
Class of columns
}
\examples{
\dontrun{
con <- dbConnect(odbc::odbc(), DSN="CMAP-MSSQL",UID="ArmLab",PWD="ArmLab2018")
#
## Choose table:
table.name <-  "tblsst_AVHRR_OI_NRT"
#
# Class of each column in the table
tbl.colClass <- getColClass(con,table.name)
print(tbl.colClass)
#
dbDisconnect(con)
}
}
