% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableFeature.R
\name{getDataSample}
\alias{getDataSample}
\title{Get the sample of size n from a table}
\usage{
getDataSample(con, table.name, n)
}
\arguments{
\item{con}{connection object to the database}

\item{table.name}{table name in the database}

\item{n}{size of sample to retrieve}
}
\value{
output data frame
}
\description{
Retrieve data sample of size n
}
\examples{
\dontrun{
con <- dbConnect(odbc::odbc(), DSN="CMAP-MSSQL",UID="ArmLab",PWD="ArmLab2018")
#
## Choose table:
table.name <-  "tblsst_AVHRR_OI_NRT"

## collect sample data
tbl.fields <- getDataSample(con,table.name,n=10)
print(tbl.fields)

dbDisconnect(con)
}
}
