% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscSubroutine.R
\name{getEqualQuery}
\alias{getEqualQuery}
\title{Create filter query with equality condition}
\usage{
getEqualQuery(range.var)
}
\arguments{
\item{range.var}{range variable}
}
\value{
list object from range variable
}
\description{
Create filter query with equality condition
}
