% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscSubroutine.R
\name{getFilterQuery}
\alias{getFilterQuery}
\title{Create filter query using specified range variable}
\usage{
getFilterQuery(range_var)
}
\arguments{
\item{range_var}{range variable}
}
\value{
list object from range variable
}
\description{
Create filter query using specified range variable
}
