% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableFeature.R
\name{getRangeNumVar}
\alias{getRangeNumVar}
\title{Get range of all the numerical variables}
\usage{
getRangeNumVar(con, table.name)
}
\arguments{
\item{con}{connection object to the database}

\item{table.name}{table name in the database}
}
\value{
table with range of numeric variables
}
\description{
Range of numeric variables
}
\examples{
\dontrun{
con <- dbConnect(odbc::odbc(), DSN="CMAP-MSSQL",UID="ArmLab",PWD="ArmLab2018")
#
# Choose table:
table.name <-  "tblsst_AVHRR_OI_NRT"
#
# Numeric variable range:
tbl.rangeNumVar <- getRangeNumVar(con,table.name)
print(tbl.rangeNumVar)
#
dbDisconnect(con)
}
}
