% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableFeature.R
\name{getSpaceTimeRange}
\alias{getSpaceTimeRange}
\title{Get space time range of a  table}
\usage{
getSpaceTimeRange(con, table.name)
}
\arguments{
\item{con}{connection object to the database}

\item{table.name}{table name in the database}
}
\value{
table with range of space time variable
}
\description{
Space time range of a table
}
\examples{
\dontrun{
con <- connect2cmap(Driver = "libtdsodbc.so")
#
## Choose table:
table.name <- "tblsst_AVHRR_OI_NRT"
#
#
# Space/time information of the table
tbl.spaceTimeInfo <- getSpaceTimeRange(con, table.name)
print(tbl.spaceTimeInfo)
#
dbDisconnect(con)
}
}
