% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableSubsetFeature.R
\name{getSubsetRangeNumVar}
\alias{getSubsetRangeNumVar}
\title{Get range of all the numerical variables of a subset}
\usage{
getSubsetRangeNumVar(con, table.name, range.var)
}
\arguments{
\item{con}{connection object to the database}

\item{table.name}{table name in the database}

\item{range.var}{range of time, latitude, longitude, depth}
}
\value{
table with range of numeric variables
}
\description{
Range of numeric variables  in a subset of table
}
\examples{
\dontrun{
con <- connect2cmap(Driver = "libtdsodbc.so")
#
## Input: Table name; variable name, space time range information
table.name <- "tblsst_AVHRR_OI_NRT" # table name
sel.var <- "sst" # choose variable
range.var <- list() # Range variable information
range.var$time <- c("2016-04-30", "2016-04-30")
range.var$lat <- c(10, 70)
range.var$lon <- c(-180, -80)
#
# Summary of the data:
tbl.subsetSummary <- getSubsetRangeNumVar(con, table.name, range.var)
print(tbl.subsetSummary)
#
dbDisconnect(con)
}
}
