% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{head}
\alias{head}
\title{Returns top records of a data set.}
\usage{
head(tableName, rows = 5, apiKey)
}
\arguments{
\item{tableName}{Name of table.}

\item{rows}{Number of rows to retrieve.}
}
\description{
Returns top records of a data set.
}
\examples{
\dontrun{
tbl_head("tblArgoMerge_REP", rows=10)
}
}
