% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorization.R
\name{initialize_cmap}
\alias{initialize_cmap}
\title{Initialize CMAP database access credentials.}
\usage{
initialize_cmap(base_url = NULL, route = NULL)
}
\arguments{
\item{base_url}{base url for the database. Default: 'https://simonscmap.com'}

\item{route}{data retrieval route of the database.  Default: "/api/data/sp?"}
}
\description{
Initialize CMAP database access credentials.
}
\examples{
\dontrun{
library(cmap4r)
# Output list with CMAP database access credentials
initialize_cmap()
}
}
