% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorization.R
\name{inti_cmap}
\alias{inti_cmap}
\title{Initialize CMAP database access credentials.}
\usage{
inti_cmap(base_url = NULL, route = NULL)
}
\arguments{
\item{base_url}{base url for the database}

\item{route}{data retrieval route of the database}
}
\description{
Initialize CMAP database access credentials.
}
\examples{
\dontrun{
library(cmap4r)
# Output list with CMAP database access credentials
inti_cmap()
}
}
