% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{is_grid}
\alias{is_grid}
\title{Returns a boolean indicating whether the variable is a gridded product or has
irregular spatial resolution.}
\usage{
is_grid(table_name, var_name)
}
\arguments{
\item{tableName}{Table name.}

\item{varName}{Variable name.}
}
\description{
Returns a boolean indicating whether the variable is a gridded product or has
irregular spatial resolution.
}
\examples{
\dontrun{
}
}
