% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscSubroutine.R
\name{match.ranges}
\alias{match.ranges}
\title{Matches range of tables.}
\usage{
match.ranges(range.tables, source)
}
\arguments{
\item{range.tables}{tables containing ranges of each ta.}

\item{source}{csv file name.}
}
\value{
names of the tables whose lat/lon/time ranges have /any/ match with
  the cruise trajectory in \code{source}.
}
\description{
Matches range of tables.
}
