% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscSubroutine.R
\name{match_ranges}
\alias{match_ranges}
\title{Matches range of tables.}
\usage{
match_ranges(range_tables, source)
}
\arguments{
\item{range_tables}{tables containing ranges.}

\item{source}{csv file name.}
}
\value{
names of the tables whose lat/lon/time ranges have /any/ match with
  the cruise trajectory in \code{source}.
}
\description{
Matches range of tables.
}
