% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalize.R
\name{matchSourceOnetable}
\alias{matchSourceOnetable}
\title{Colocalization of a dataset using lattitude, longitude, depth and time information}
\usage{
matchSourceOnetable(con, source, table.name, sel.vars, latMargin,
  lonMargin, timeMargin, depthRange = NULL, orderby = NULL,
  N = 1e+05)
}
\arguments{
\item{con}{connection object to the database}

\item{source}{The file name of the csv source file that contains the
lat/lon/date triplets.}

\item{table.name}{table name.}

\item{sel.vars}{variables in this table that are to be queried.}

\item{latMargin}{latitude margin.}

\item{lonMargin}{longitude margin.}

\item{timeMargin}{time margin.}

\item{depthRange}{range of depth}

\item{orderby}{orderby variable in the selected table}

\item{N}{description to add later}
}
\value{
Data frame containing the colocalized data. Each row is an original
  time/lat/lon.
}
\description{
Takes all the lat/lon/time triplets in the comma separated \code{source}
file, collects and aggregates (mean and standard deviation) each variable's
data in a rectangle around each triplet. The size of the rectangle is
determined by a user-input margin in lat, lon and time.
}
\examples{
\dontrun{
## Setup
con = DBI::dbConnect(odbc::odbc(), DSN="CMAP-MSSQL",UID="ArmLab",PWD="ArmLab2018")
latMargin = 0.3
lonMargin = 0.3
timeMargin = 1
source = 'somefile.csv'
table.name = "tblSST_AVHRR_OI_NRT"
sel.var = "sst"
res = matchSourceOneTable(con, source, table.name, sel.var, latMargin, lonMargin, timeMargin)
print(head(res))
}
}
