% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscSubroutine.R
\name{one_overlap}
\alias{one_overlap}
\title{Helper that checks: (StartA <= EndB) and (EndA >= StartB), in 1d.}
\usage{
one_overlap(min1, max1, min2, max2)
}
\arguments{
\item{min1}{minimum of first range.}

\item{max1}{maximum of first range.}

\item{min2}{minimum of second range.}

\item{max2}{maximum of second range.}
}
\value{
\code{TRUE} if there is overlap.
}
\description{
Helper that checks: (StartA <= EndB) and (EndA >= StartB), in 1d.
}
