% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscSubroutine.R
\name{only_has_date}
\alias{only_has_date}
\title{Takes in date (can be a string) and returns TRUE of only date exists, and no
time information.}
\usage{
only_has_date(mydate)
}
\arguments{
\item{mydate}{Date}
}
\value{
\code{TRUE} if the time-of-day information does not exist.
}
\description{
Takes in date (can be a string) and returns TRUE of only date exists, and no
time information.
}
