% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cmap.R
\name{plot_hist}
\alias{plot_hist}
\title{Plot histogram of the variable of interest}
\usage{
plot_hist(table_name, sel_var, range_var, type = c("plotly",
  "ggplot")[1], export_data = FALSE)
}
\arguments{
\item{table_name}{name of the table from CMAP}

\item{sel_var}{select a variable in the table}

\item{range_var}{range variable}

\item{type}{choose a type of plot object: 'plotly','ggplot}

\item{export_data}{boolean variable to export data or not. The data will be saved in the working directory.}
}
\value{
plot object of the chosen type
}
\description{
Histogram object of chosen type
}
\examples{
\donttest{
library(cmap4r)
#
# Inpit variable:
table_list <- c("tblSST_AVHRR_OI_NRT", "tblArgoMerge_REP", "tblArgoMerge_REP")
var_list <- c("sst", "argo_merge_temperature_adj", "argo_merge_salinity_adj")
#
# variable "sst" selected from "tblSST_AVHRR_OI_NRT"
selIndex <- 1
table_name <- table_list[selIndex]
sel_var <- var_list[selIndex]
#
#
# Range variable
range_var <- list()
range_var$lat <- c(20, 24)
range_var$lon <- c(-170, 150)
range_var$time <- c("2016-04-30", "2016-04-30")
#
#
# Plot function available for R User
p <- plot_hist(table_name, sel_var, range_var, "ggplot")
p <- plot_hist(table_name, sel_var, range_var, "plotly")
p
#
}
}
