% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cmap.R
\name{plot_ts}
\alias{plot_ts}
\title{Plot variable of interest with varying time.}
\usage{
plot_ts(table_name, sel_var, range_var, type = c("plotly", "ggplot")[1],
  export_data = FALSE)
}
\arguments{
\item{table_name}{table with depth and variable of interest}

\item{sel_var}{select a variable in the table}

\item{range_var}{range variable}

\item{type}{choose a type of plot object: 'plotly','ggplot}

\item{export_data}{boolean variable to export data or not. The data will be saved in the working directory.}
}
\value{
plot object of the chosen type
}
\description{
Times eries plot object.
}
\examples{
\donttest{
library(cmap4r)
#
# Input variable:
table_list <- c("tblSST_AVHRR_OI_NRT", "tblAltimetry_REP", "tblPisces_NRT")
var_list <- c("sst", "sla", "NO3")
#
# selected "sst" from the table "tblSST_AVHRR_OI_NRT"
selIndex <- 1
table_name <- table_list[selIndex]
sel_var <- var_list[selIndex]
#
# # Example I:
range_var <- list()
range_var$lat <- c(25, 30)
range_var$lon <- c(-160, -155)
range_var$time <- c("2016-03-29", "2016-05-29")
#
# Plot -- Time series:
p <- plot_ts(table_name, sel_var, range_var)
p
#
}
}
