% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cmap.R
\name{plot_xy}
\alias{plot_xy}
\title{XY plot}
\usage{
plot_xy(table_list, var_list, lat1, lat2, lon1, lon2, dt1, dt2,
  depth1 = NULL, depth2 = NULL, agg_var, type = c("plotly",
  "ggplot")[1], export_data = FALSE)
}
\arguments{
\item{table_list}{list of  tables}

\item{var_list}{list of  corresponding variable}

\item{lat1}{start latitude [degree N].}

\item{lat2}{end latitude [degree N].}

\item{lon1}{start longitude [degree E].}

\item{lon2}{end longitude [degree E].}

\item{dt1}{start date or datetime.}

\item{dt2}{end date or datetime.}

\item{depth1}{start depth [m].}

\item{depth2}{end depth [m].}

\item{agg_var}{aggregate variable}

\item{type}{choose a type of plot object: 'plotly','ggplot}

\item{export_data}{boolean variable to export data or not. The data will be saved in the working directory.}
}
\value{
plot object of the chosen type
}
\description{
Scatter plot of two variable of interest
}
\examples{
\donttest{

# Inpit variable:
#
table_list <- c("tblSST_AVHRR_OI_NRT", "tblAltimetry_REP")
var_list <- c("sst", "sla")
#
# Range variable [lat,lon,time,depth]
lat1 = 25; lat2 = 30
lon1 = -160; lon2 = -155
dt1 =  "2016-03-29"; dt2 = "2016-05-29"
depth1 <- NULL; depth2 =  NULL

#
agg_var <- "time"

# xy plot
out <- plot_xy(table_list, var_list, lat1, lat2, lon1, lon2,
               dt1, dt2, depth1, depth2, agg_var)
out
}
}
