% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{process_response_to_tibble}
\alias{process_response_to_tibble}
\title{Helper to get response to tibble.}
\usage{
process_response_to_tibble(response, route)
}
\description{
Helper to get response to tibble.
}
