% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{query}
\alias{query}
\title{Takes a custom query, issues a request to the API, and returns the results
in form of a dataframe.}
\usage{
query(myquery, apiKey)
}
\arguments{
\item{myquery}{An "EXEC ..." string.}

\item{apiKey}{The API Key.}
}
\description{
Takes a custom query, issues a request to the API, and returns the results
in form of a dataframe.
}
