% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscSubroutine.R
\name{range2list}
\alias{range2list}
\title{Convert range object to list object: It will be used latter for creating query.}
\usage{
range2list(range.var)
}
\arguments{
\item{range.var}{range variable}
}
\value{
list object from range variable
}
\description{
Convert range object to list object: It will be used latter for creating query.
}
