% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscSubroutine.R
\name{range_tibble_to_table}
\alias{range_tibble_to_table}
\title{Helper to take the tibbles containing min/maxs of the variables, and
transforms it into a set of 2 by #(variables) data frames.}
\usage{
range_tibble_to_table(myrange)
}
\arguments{
\item{myrange}{Result of \code{return_range()}, containing user-desired
ranges of variables.}
}
\value{
2 by #(variables) data frame.
}
\description{
Helper to take the tibbles containing min/maxs of the variables, and
transforms it into a set of 2 by #(variables) data frames.
}
