% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscSubroutine.R
\name{rangeVarCheck}
\alias{rangeVarCheck}
\title{Check if the specified range variable is present in the table or not}
\usage{
rangeVarCheck(con, table_name, range_var)
}
\arguments{
\item{con}{connection object to the database}

\item{table_name}{table name in the database}

\item{range_var}{range variable}
}
\value{
0 or index of range variable
}
\description{
Check if the specified range variable is present in the table or not
}
