% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{request}
\alias{request}
\title{Send API request. Going from (payload -> response -> df). Near-identical
function in Python.}
\usage{
request(payload, route, apiKey)
}
\description{
Send API request. Going from (payload -> response -> df). Near-identical
function in Python.
}
