% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscSubroutine.R
\name{return_range}
\alias{return_range}
\title{Helper for matchSource() that takes in a lat/lon/date triplet and the
margins of error for query, and returns a "range" variable.}
\usage{
return_range(lat, lon, dt, latMargin, lonMargin, timeMargin, depthRange)
}
\arguments{
\item{lat}{center latitude.}

\item{lon}{center longitude.}

\item{dt}{center date.}

\item{latMargin}{margin for latitude, around the center.}

\item{lonMargin}{margin for longitude, around the center.}

\item{timeMargin}{margin for time, around the center (\code{dt}).}

\item{depthRange}{Range of depth of interest}
}
\value{
A list containing the desired lat/lon/time ranges.
}
\description{
Helper for matchSource() that takes in a lat/lon/date triplet and the
margins of error for query, and returns a "range" variable.
}
