% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscSubroutine.R
\name{return_range_from_all_triplets}
\alias{return_range_from_all_triplets}
\title{Helper for matchSource() that takes in many lat/lon/date triplets and the
margins of error for query, and returns a "range" variable for all triplets
(with margin).}
\usage{
return_range_from_all_triplets(source_table, latMargin, lonMargin,
  timeMargin, depthRange)
}
\arguments{
\item{source_table}{a matrix with 3 columns which each are: lat (center
latitude), lon (center longitude) dt (center date).}

\item{latMargin}{margin for latitude, around the center.}

\item{lonMargin}{margin for longitude, around the center.}

\item{timeMargin}{margin for time, around the center (\code{dt}).}

\item{depthRange}{Range of depth of interest}
}
\value{
A list containing the desired lat/lon/time ranges.
}
\description{
Helper for matchSource() that takes in many lat/lon/date triplets and the
margins of error for query, and returns a "range" variable for all triplets
(with margin).
}
