% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{section}
\alias{section}
\title{Returns a subset of data according to space-time constraints.  The results
are ordered by time, lat, lon, and depth.}
\usage{
section(table, variable, dt1, dt2, lat1, lat2, lon1, lon2, depth1, depth2)
}
\description{
Returns a subset of data according to space-time constraints.  The results
are ordered by time, lat, lon, and depth.
}
\examples{
\dontrun{
library(cmap4r)
table <- "tblsst_AVHRR_OI_NRT"
variable <- "sst"
lat1 = 10
lat2 = 70
lon1 = -180
lon2 = -80
dt1 = "2016-04-30"
dt2 = "2016-04-30"
depth1 = 0
depth2 = 30
section(table, variable, dt1, dt2, lat1, lat2, lon1, lon2, depth1, depth2)
}

}
