% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorization.R
\name{set_authorization}
\alias{set_authorization}
\title{Set/save the user API key for accessing the CMAP database.}
\usage{
set_authorization(reset = FALSE)
}
\arguments{
\item{reset}{boolean variable for set authorization key for API access to the database}
}
\description{
Set/save the user API key for accessing the CMAP database.
}
\examples{
\dontrun{

# To set the API authorization key
set_authorization()

# To reset the authorization key
set_authorization(reset = TRUE)
}
}
