% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{space_time}
\alias{space_time}
\title{Returns a subset of data according to
space-time constraints.  The results are ordered by time, lat, lon, and
depth (if exists).}
\usage{
space_time(table, variable, dt1, dt2, lat1, lat2, lon1, lon2, depth1,
  depth2)
}
\arguments{
\item{table_name}{table name in the database}

\item{sel_var}{select a variable in the table}

\item{range_var}{range of time, latitude, longitude, depth}
}
\value{
subset of a table as data frame
}
\description{
Get subset of a table as data frame
}
\examples{
\dontrun{
library(cmap4r)
#
## Input: Table name; variable name, space time range information
table_name <- "tblsst_AVHRR_OI_NRT" # table name
sel_var <- "sst" # choose variable
range_var <- list() # Range variable [lat,lon,time]
range_var$lat <- c(10, 70)
range_var$lon <- c(-180, -80)
range_var$time <- c("2016-04-30", "2016-04-30")
#
## Subset selection:
tbl.subset <- get_spacetime(table_name, sel_var, range_var)
head(tbl.subset)
#
}
}
