% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{stored_proc}
\alias{stored_proc}
\title{Executes a stored procedure. (Near-identical function in python.) Executes a
stored-procedure and returns the results in form of a dataframe. Similar to
\code{query()}.}
\usage{
stored_proc(query, args, apiKey)
}
\description{
Executes a stored procedure. (Near-identical function in python.) Executes a
stored-procedure and returns the results in form of a dataframe. Similar to
\code{query()}.
}
