% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{subset}
\alias{subset}
\title{Returns a subset of data according to space-time constraints. (Identical function in python.)}
\usage{
subset(spName, table, variable, dt1, dt2, lat1, lat2, lon1, lon2, depth1,
  depth2, apiKey)
}
\description{
Returns a subset of data according to space-time constraints. (Identical function in python.)
}
