% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscSubroutine.R
\name{subset_from_tbl}
\alias{subset_from_tbl}
\title{Subsets a subtble from an OFFLINE table (3 columns called "lat" "lon" and
"time").}
\usage{
subset_from_tbl(tbl, range_var)
}
\arguments{
\item{tbl}{3 column matrix.}

\item{range_var}{Data frame containing information about the "rectangle" of
interest.}
}
\value{
3 column matrix with only a subset of the table.
}
\description{
Subsets a subtble from an OFFLINE table (3 columns called "lat" "lon" and
"time").
}
