% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscSubroutine.R
\name{tableVarMatch}
\alias{tableVarMatch}
\title{Check if the specified variables are present in the table or not}
\usage{
tableVarMatch(con, table_name, varname)
}
\arguments{
\item{con}{connection object to the database}

\item{table_name}{table name in the database}

\item{varname}{names of variables}
}
\value{
0 or index of range variable
}
\description{
Check if the specified variables are present in the table or not
}
