% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableFeature.R
\name{tbl_nobs}
\alias{tbl_nobs}
\title{Get the number of observations in the table}
\usage{
tbl_nobs(con, table_name, type = c("approx", "first-last", "exact")[1],
  idvar = "ID")
}
\arguments{
\item{con}{connection object to the database}

\item{table_name}{table name in the database}

\item{type}{method of getting number of observation: a) "approx" - fast retrieval but inaccurate;
b) "first-last"- using ID of first and last observation; c) "exact" - use COUNT_BIG to get exact count}

\item{idvar}{ID variable used in "first-last" method}
}
\value{
number of observation in the table
}
\description{
Number of observations
}
\examples{
\dontrun{
library(cmap4r)
con <- connect_cmap(Driver = "libtdsodbc.so")
#
## Choose table:
table_name <- "tblsst_AVHRR_OI_NRT"

# Number of observations
nObs <- tbl_nobs(con, table_name)
nObs
#
dbDisconnect(con)
}
}
