% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableFeature.R
\name{tbl_numvar_range}
\alias{tbl_numvar_range}
\title{Get range of all the numerical variables}
\usage{
tbl_numvar_range(con, table.name)
}
\arguments{
\item{con}{connection object to the database}

\item{table.name}{table name in the database}
}
\value{
table with range of numeric variables
}
\description{
Range of numeric variables
}
\examples{
\dontrun{
library(cmap4r)
con <- connect_cmap(Driver = "libtdsodbc.so")
#
# Choose table:
table.name <- "tblsst_AVHRR_OI_NRT"
#
# Numeric variable range:
tbl.rangeNumVar <- tbl_numvar_range(con, table.name)
print(tbl.rangeNumVar)
#
dbDisconnect(con)
}
}
