% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableFeature.R
\name{tbl_sample}
\alias{tbl_sample}
\title{Get the sample of size n from a table}
\usage{
tbl_sample(con, table.name, n)
}
\arguments{
\item{con}{connection object to the database}

\item{table.name}{table name in the database}

\item{n}{size of sample to retrieve}
}
\value{
output data frame
}
\description{
Retrieve data sample of size n
}
\examples{
\dontrun{
library(cmap4r)
con <- connect_cmap(Driver = "libtdsodbc.so")
#
## Choose table:
table.name <- "tblsst_AVHRR_OI_NRT"

## collect sample data
tbl.fields <- tbl_sample(con, table.name, n = 10)
print(tbl.fields)

dbDisconnect(con)
}
}
