% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableFeature.R
\name{tbl_vartype}
\alias{tbl_vartype}
\title{Get class of columns in the table}
\usage{
tbl_vartype(con, table.name)
}
\arguments{
\item{con}{connection object to the database}

\item{table.name}{table name in the database}
}
\value{
column name with its class
}
\description{
Class of columns
}
\examples{
\dontrun{
library(cmap4r)
con <- connect_cmap(Driver = "libtdsodbc.so")
#
## Choose table:
table.name <- "tblsst_AVHRR_OI_NRT"
#
# Class of each column in the table
tbl.colClass <- tbl_vartype(con, table.name)
print(tbl.colClass)
#
dbDisconnect(con)
}
}
