% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{time_series}
\alias{time_series}
\title{Retrieve timeseries data from  CMAP using space time range parameters}
\usage{
time_series(table, variable, dt1, dt2, lat1, lat2, lon1, lon2, depth1,
  depth2, interval = NULL)
}
\arguments{
\item{interval}{choose time series data aggregation at [weekly, montly, quaterly, annual, none]}
}
\value{
subset of a table as data frame
}
\description{
Get subset of a table as data frame
}
\examples{
\dontrun{
library(cmap4r)
#
table <- "tblHOT_Bottle" # table name
variable <- "SiO4_bottle_hot" # choose variable
lat1 = 22
lat2 = 23
lon1 = -159
lon2 = -157
dt1 = 1988-12-01
dt2 = 2016-10-15
time_series(table, variable, dt1, dt2, lat1, lat2, lon1, lon2, depth1, depth2) ## not complete
#
}
}
