% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{uspInterval_indicator}
\alias{uspInterval_indicator}
\title{Retrieve stored procedure for extracting timeseries data aggregated at
weekely, monthly, quaterly, annual or none level.}
\usage{
uspInterval_indicator(interval)
}
\description{
Retrieve stored procedure for extracting timeseries data aggregated at
weekely, monthly, quaterly, annual or none level.
}
