% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{validate_sp_args}
\alias{validate_sp_args}
\title{Validating the format of sp arguments.}
\usage{
validate_sp_args(table, variable, dt1, dt2, lat1, lat2, lon1, lon2, depth1,
  depth2)
}
\arguments{
\item{table}{(string) table name (each data set is stored in one or more
than one table).}

\item{variable}{(string) variable short name which directly corresponds to a
field name in the table.}

\item{dt1}{(string of format "YYYY-MM-DDTHH:MM:SS" start date (or datetime
simply as YYYY-MM-DD). Caution!! This function doesn't check for the exact
format of the input, and only checks for whether it is a string. Same for
\code{dt2}.}

\item{dt2}{(string of format "YYYY-MM-DDTHH:MM:SS" end date (or datetime,
simply as YYYY-MM_DD).}

\item{lat1}{(Numeric) start latitude [degree N].}

\item{lat2}{(Numeric) end latitude [degree N].}

\item{lon1}{(Numeric) start longitude [degree E].}

\item{lon2}{(Numeric) end longitude [degree E].}

\item{depth1}{(Numeric) start depth [m].}

\item{depth2}{(Numeric) end depth [m].}
}
\description{
Validating the format of sp arguments.
}
