% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cruise_info.R
\name{cruise_bounds}
\alias{cruise_bounds}
\title{Returns a dataframe containing cruise boundaries in space and time.}
\usage{
cruise_bounds(cruisename)
}
\arguments{
\item{cruisename}{Name of cruise.}
}
\description{
Returns a dataframe containing cruise boundaries in space and time.
}
\examples{
\donttest{

#
## Variable attribute:
cruiseBounds <- cruise_bounds('KOK1606')
#
}
}
