% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{cruise_by_name}
\alias{cruise_by_name}
\title{Returns a dataframe containing cruise info using cruise name.}
\usage{
cruise_by_name(cruisename)
}
\arguments{
\item{cruisename}{Name of cruise.}
}
\description{
Returns a dataframe containing cruise info using cruise name.
}
