% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cruise_info.R
\name{cruise_trajectory}
\alias{cruise_trajectory}
\title{Returns a dataframe containing the cruise trajectory (Identical function in python)}
\usage{
cruise_trajectory(cruisename)
}
\arguments{
\item{cruisename}{name of cruise.}
}
\description{
Returns a dataframe containing the cruise trajectory (Identical function in python)
}
\examples{
\donttest{
library(cmap4r)
#
## Variable attribute:
cruiseTrajectory  <- cruise_trajectory('KM1513')
#
}
}
