% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{cruise_variables}
\alias{cruise_variables}
\title{Returns a dataframe containing all registered variables (at Simons CMAP) during a cruise.}
\usage{
cruise_variables(cruiseName)
}
\description{
Returns a dataframe containing all registered variables (at Simons CMAP) during a cruise.
}
