% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_info.R
\name{get_catalog}
\alias{get_catalog}
\title{Returns a catalog of the Simons CMAP database.}
\usage{
get_catalog()
}
\value{
Simons CMAP catalog as dataframe object
}
\description{
Returns a catalog of the Simons CMAP database.
}
\examples{
\donttest{
#
## Variable attribute:
cmap.catalog <- get_catalog()
cmap.catalog
#
}
}
