% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_info.R
\name{get_columns}
\alias{get_columns}
\title{Returns the list of column variables in a table.}
\usage{
get_columns(tableName)
}
\arguments{
\item{tableName}{table name from the Simons CMAP database. Use "get_catalog()" to retrieve list of tables on the database.}
}
\value{
column variables name of a table as dataframe
}
\description{
Returns the list of column variables in a table.
}
\examples{
\donttest{
#
## Input: Table name;
tableName <- "tblAMT13_Chisholm" # table name
#
## Subset selection:
tbl.columns <- get_columns(tableName)
tbl.columns
#
}
}
