% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_depthprofile}
\alias{get_depthprofile}
\title{Retrieve depth profile data from  CMAP using space time range parameters}
\usage{
get_depthprofile(table_name, sel_var, lat1, lat2, lon1, lon2, dt1, dt2,
  depth1, depth2)
}
\arguments{
\item{table_name}{table name in the database}

\item{sel_var}{select a variable in the table}

\item{lat1}{start latitude [degree N].}

\item{lat2}{end latitude [degree N].}

\item{lon1}{start longitude [degree E].}

\item{lon2}{end longitude [degree E].}

\item{dt1}{start date or datetime.}

\item{dt2}{end date or datetime.}

\item{depth1}{start depth [m].}

\item{depth2}{end depth [m].}
}
\value{
subset of a table as data frame
}
\description{
Get subset of a table as data frame
}
\examples{
\dontrun{
library(cmap4r)
#
## Input: Table name; variable name, space time range information
table_name <- "tblArgoMerge_REP" # table name
sel_var <- "argo_merge_chl_adj" # choose variable
# Range variable [lat,lon,time]
lat1 = 20; lat2 = 24
lon1 = -170; lon2 = -150
dt1 = "2016-04-30"; dt2 = "2016-04-30"
depth1 <- 0; depth2 =  1500
#
## Subset selection:
tbl.subset <- get_depthprofile(table_name, sel_var, lat1, lat2, lon1, lon2,
                               dt1, dt2, depth1, depth2)
head(tbl.subset)
#
}
}
