% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_info.R
\name{get_head}
\alias{get_head}
\title{Returns top  n records  from a table on CMAP}
\usage{
get_head(tableName, nrows = 5)
}
\arguments{
\item{tableName}{name of table.}

\item{nrows}{number of rows to retrieve.}
}
\value{
subset of a table as data frame
}
\description{
Returns top  n records  from a table on CMAP
}
\examples{
\donttest{
library(cmap4r)

## Input: Table name;
table_name <- "tblArgoMerge_REP" # table name
#
## Top n rows:
tbl.subset <- get_head(table_name, nrows=10)
tbl.subset
}
}
