% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_info.R
\name{get_metadata}
\alias{get_metadata}
\title{Returns a dataframe containing the associated metadata. The inputs can be strings (if only one table, and variable is passed) or a list of string literals.}
\usage{
get_metadata(tables, variables)
}
\arguments{
\item{tables}{vector of table names.}

\item{variables}{vector of variable names in the tables.}
}
\description{
Returns a dataframe containing the associated metadata. The inputs can be strings (if only one table, and variable is passed) or a list of string literals.
}
\examples{
\dontrun{
library(cmap4r)
#
## Input:
tables <- c('tblsst_AVHRR_OI_NRT', 'tblArgoMerge_REP') # table name
variables <- c('sst', 'argo_merge_salinity_adj') # variable name

metadata <- get_metadata(tables, variables)
metadata
}
}
