% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{get_metadata_noref}
\alias{get_metadata_noref}
\title{Returns a dataframe containing the associated metadata of a variable (such
as data source, distributor, references, and etc..). The inputs can be
string literals (if only one table, and variable is passed) or a list of
string literals.}
\usage{
get_metadata_noref(table, variable)
}
\arguments{
\item{table}{Table name.}

\item{variable}{Variable name.}
}
\value{
Metadata, in the form of a data frame.
}
\description{
Returns a dataframe containing the associated metadata of a variable (such
as data source, distributor, references, and etc..). The inputs can be
string literals (if only one table, and variable is passed) or a list of
string literals.
}
