% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{get_references}
\alias{get_references}
\title{Returns a dataframe containing references associated with a data set.}
\usage{
get_references(datasetID)
}
\arguments{
\item{datasetID}{ID of dataset.}
}
\description{
Returns a dataframe containing references associated with a data set.
}
