% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_section}
\alias{get_section}
\title{Retrieve subset of a table from CMAP using space time range parameters}
\usage{
get_section(table_name, sel_var, range_var)
}
\arguments{
\item{table_name}{table name in the database}

\item{sel_var}{select a variable in the table}

\item{range_var}{range of time, latitude, longitude, depth}
}
\value{
subset of a table as data frame
}
\description{
Get subset of a table as data frame
}
\examples{
\donttest{
library(cmap4r)
#
## Input: Table name; variable name, space time range information
table_name <- "tblPisces_NRT" # table name
sel_var <- "NO3" # choose variable
range_var <- list() # Range variable [lat,lon,time]
range_var$lat <- c(10, 60)
range_var$lon <- c(-160, -158)
range_var$time <- c("2016-04-30", "2016-04-30")
range_var$depth <- c(0, 5000)
#
## Subset selection:
tbl.subset <- get_section(table_name, sel_var, range_var)
head(tbl.subset)
#
}
}
