% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_spacetime}
\alias{get_spacetime}
\title{Returns a subset of data according to
space-time constraints.  The results are ordered by time, lat, lon, and
depth (if exists).}
\usage{
get_spacetime(table_name, sel_var, lat1, lat2, lon1, lon2, dt1, dt2,
  depth1 = NULL, depth2 = NULL)
}
\arguments{
\item{table_name}{table name in the database}

\item{sel_var}{select a variable in the table}

\item{lat1}{start latitude [degree N].}

\item{lat2}{end latitude [degree N].}

\item{lon1}{start longitude [degree E].}

\item{lon2}{end longitude [degree E].}

\item{dt1}{start date or datetime.}

\item{dt2}{end date or datetime.}

\item{depth1}{start depth [m].}

\item{depth2}{end depth [m].}
}
\value{
subset of a table as data frame
}
\description{
Get subset of a table as data frame
}
\examples{
\donttest{
library(cmap4r)
#
## Input: Table name; variable name, space time range information
table_name <- "tblsst_AVHRR_OI_NRT" # table name
sel_var <- "sst" # choose variable
# Range variable [lat,lon,time]
lat1 = 10; lat2 = 70
lon1 = -180; lon2 = -80
dt1 = "2016-04-30"; dt2 = "2016-04-30"
#
## Subset selection:
tbl.subset <- get_spacetime(table_name, sel_var, lat1, lat2,
                               lon1, lon2, dt1, dt2)
head(tbl.subset)
#
}
}
