% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_spacetime}
\alias{get_spacetime}
\title{Returns a subset of table using the specified space-time range inputs.}
\usage{
get_spacetime(tableName, varName, lat1, lat2, lon1, lon2, dt1, dt2,
  depth1 = NULL, depth2 = NULL)
}
\arguments{
\item{tableName}{table name from the Simons CMAP database. Use "get_catalog()" to retrieve list of tables on the database.}

\item{varName}{specify short name of a variable in the table. Pass the input "*" to retrieve all fields in the table.  Use "get_catalog()" to retrieve list of table variables on the database.}

\item{lat1}{start latitude [degree N] of the meridional cut; ranges from -90° to 90°.}

\item{lat2}{end latitude [degree N] of the meridional cut; ranges from -90° to 90°.}

\item{lon1}{start longitude [degree E]  of the zonal cut; ranges from  -180° to 180°.}

\item{lon2}{end longitude [degree E] of the zonal cut; ranges from  -180° to 180°.}

\item{dt1}{start date or datetime (lower bound of temporal cut). Example values: '2016-05-25' or '2017-12-10 17:25:00'}

\item{dt2}{end date or datetime (upper bound of temporal cut). Example values: '2016-04-30' or '2016-04-30 17:25:00'}

\item{depth1}{positive value specifying the start depth [m] of the vertical cut. Note that depth  is 0 at surface and grows towards ocean floor.}

\item{depth2}{positive value specifying the end depth [m]of the vertical cut. Note that depth  is 0 at surface and grows towards ocean floor.}
}
\value{
required subset of the table is ordered by time, lat, lon, and depth (if exists).
}
\description{
Returns a subset of table using the specified space-time range inputs.
}
\examples{
\donttest{
## Input: Table name; variable name, space time range information
tableName <- "tblsst_AVHRR_OI_NRT" # table name
varName <- "sst" # choose variable
# Range variable [lat,lon,time]
lat1 = 10; lat2 = 70
lon1 = -180; lon2 = -80
dt1 = "2016-04-30"; dt2 = "2016-04-30"
#
## Subset selection:
tbl.subset <- get_spacetime(tableName, varName, lat1, lat2,
                               lon1, lon2, dt1, dt2)
head(tbl.subset)
#
}
}
