% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_timeseries}
\alias{get_timeseries}
\title{Retrieve the subset of a table from the Simons CMAP databse using the space-time range inputs (dt1, dt2, lat1, lat2, lon1, lon2, depth1, depth2). The output is aggregated by time, i.e., we compute the mean and standard deviation of a variable of interest in the specified range of inputs. Also, the function is capable of binning the timeseries data weekly, monthly, quarterly, or annually.}
\usage{
get_timeseries(tableName, varName, lat1, lat2, lon1, lon2, dt1, dt2,
  depth1 = NULL, depth2 = NULL, interval = NULL)
}
\arguments{
\item{tableName}{table name from the Simons CMAP database. Use "get_catalog()" to retrieve list of tables on the database.}

\item{varName}{specify short name of a variable in the table.  Use "get_catalog()" to retrieve list of table variables on the database.}

\item{lat1}{start latitude [degree N] of the meridional cut; ranges from -90° to 90°.}

\item{lat2}{end latitude [degree N] of the meridional cut; ranges from -90° to 90°.}

\item{lon1}{start longitude [degree E]  of the zonal cut; ranges from  -180° to 180°.}

\item{lon2}{end longitude [degree E] of the zonal cut; ranges from  -180° to 180°.}

\item{dt1}{start date or datetime (lower bound of temporal cut). Example values: '2016-05-25' or '2017-12-10 17:25:00'}

\item{dt2}{end date or datetime (upper bound of temporal cut). Example values: '2016-04-30' or '2016-04-30 17:25:00'}

\item{depth1}{positive value specifying the start depth [m] of the vertical cut. Note that depth  is 0 at surface and grows towards ocean floor.}

\item{depth2}{positive value specifying the end depth [m]of the vertical cut. Note that depth  is 0 at surface and grows towards ocean floor.}

\item{interval}{choose time series data aggregation at [weekly, montly, quaterly, annual, none]. Default none.}
}
\value{
required aggregated output is ordered by time, lat, lon, and depth (if exists).
}
\description{
Retrieve the subset of a table from the Simons CMAP databse using the space-time range inputs (dt1, dt2, lat1, lat2, lon1, lon2, depth1, depth2). The output is aggregated by time, i.e., we compute the mean and standard deviation of a variable of interest in the specified range of inputs. Also, the function is capable of binning the timeseries data weekly, monthly, quarterly, or annually.
}
\examples{
\dontrun{
## Input: Table name; variable name, space time range information
tableName <- "tblHOT_Bottle" # table name
varName <- "SiO4_bottle_hot" # choose variable
# Range variable [lat,lon,time]
lat1 = 22; lat2 = 23
lon1 = -159; lon2 = -157
dt1 = "1988-12-01"; dt2 = "2016-10-15"

# Subset selection:
tbl.subset <- get_timeseries(tableName, varName, lat1, lat2,
                                lon1, lon2, dt1, dt2)
head(tbl.subset)
#
}
}
