% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R, R/table_info.R
\name{get_var}
\alias{get_var}
\title{Returns a single-row dataframe from tblVariables containing info associated with varName.
This method is mean to be used internally and will not be exposed at documentations.}
\usage{
get_var(tableName, varName)

get_var(tableName, varName)
}
\arguments{
\item{tableName}{Table name.}

\item{varName}{Variable name.}

\item{tableName}{table name.}

\item{varName}{variable name.}
}
\description{
This method is mean to be used internally and will not be exposed at documentations.
}
\examples{
\dontrun{
library(cmap4r)
#
## Input: Table name; variable name
table_name <- "tblArgoMerge_REP" # table name
var_name <- "argo_merge_chl_adj" # variable name
#
## Variable attribute:
tbl.var <- get_var(table_name, var_name)
}
\donttest{
library(cmap4r)
#
## Input: Table name; variable name
table_name <- "tblArgoMerge_REP" # table name
var_name <- "argo_merge_chl_adj" # variable name
#
## Variable attribute:
tbl.var <- get_var(table_name, var_name)
tbl.var
}
}
