% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{get_var}
\alias{get_var}
\title{Returns a single-row dataframe from tblVariables containing info associated with varName.
This method is mean to be used internally and will not be exposed at documentations.}
\usage{
get_var(tableName, varName, apiKey)
}
\arguments{
\item{tableName}{Table name.}

\item{varName}{Variable name.}
}
\description{
Returns a single-row dataframe from tblVariables containing info associated with varName.
This method is mean to be used internally and will not be exposed at documentations.
}
