% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_info.R
\name{get_var_catalog}
\alias{get_var_catalog}
\title{Return a single-row dataframe about a table variable from the catalog of the Simons CMAP database.}
\usage{
get_var_catalog(tableName, varName)
}
\arguments{
\item{tableName}{table name from the Simons CMAP database. Use "get_catalog()" to retrieve list of tables on the database.}

\item{varName}{specify short name of a variable in the table.}
}
\value{
attributes of variable on the catalog as dataframe.
}
\description{
Return a single-row dataframe about a table variable from the catalog of the Simons CMAP database.
}
\examples{
\donttest{

#
## Input: Table name; variable name
tableName <- "tblArgoMerge_REP" # table name
varName <- "argo_merge_chl_adj" # variable name
#
## Variable attribute:
tbl.catlog.var <- get_var_catalog(tableName, varName)
tbl.catlog.var
#
}
}
