% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R, R/table_info.R
\name{get_var_catalog}
\alias{get_var_catalog}
\title{Returns a single-row dataframe from catalog (udfCatalog) containing all of
the variable's info at catalog.}
\usage{
get_var_catalog(tableName, varName)

get_var_catalog(tableName, varName)
}
\arguments{
\item{tableName}{Table name.}

\item{varName}{Variable name.}

\item{tableName}{table name.}

\item{varName}{variable name.}
}
\value{
Single-row data frame.

Single-row data frame.
}
\description{
Returns a single-row dataframe from catalog (udfCatalog) containing all of
the variable's info at catalog.

Returns a single-row dataframe from catalog (udfCatalog) containing all of
the variable's info at catalog.
}
\examples{
\dontrun{
tableName <- "tblArgoMerge_REP" # table name
varName <- "argo_merge_chl_adj" # variable name
get_var_catalog(tableName, varName)
}
\donttest{
library(cmap4r)
#
## Input: Table name; variable name
table_name <- "tblArgoMerge_REP" # table name
var_name <- "argo_merge_chl_adj" # variable name
#
## Variable attribute:
tbl.catlog.var <- get_var_catalog(table_name, var_name)
tbl.catlog.var
#
}
}
