% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{get_var_long_name}
\alias{get_var_long_name}
\title{Returns the long name of a given variable.}
\usage{
get_var_long_name(tableName, varName)
}
\arguments{
\item{tableName}{Table name.}

\item{varName}{Variable name.}
}
\description{
Returns the long name of a given variable.
}
\examples{
\dontrun{
tableName <- "tblArgoMerge_REP" # table name
varName <- "argo_merge_chl_adj" # variable name
get_var_long_name(tableName, varName)
}
}
