% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_info.R
\name{get_var_long_name}
\alias{get_var_long_name}
\title{Returns the long name of a given variable.}
\usage{
get_var_long_name(tableName, varName)
}
\arguments{
\item{tableName}{table name.}

\item{varName}{variable name.}
}
\description{
Returns the long name of a given variable.
}
\examples{
\donttest{
tableName <- "tblArgoMerge_REP" # table name
varName <- "argo_merge_chl_adj" # variable name
#
## Variable attribute:
varLonName <- get_var_long_name(tableName, varName)
varLonName
#
}
}
