% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R, R/table_info.R
\name{get_var_resolution}
\alias{get_var_resolution}
\title{Returns a single-row dataframe from catalog (udfCatalog) containing the
variable's spatial and temporal resolutions.}
\usage{
get_var_resolution(tableName, varName)

get_var_resolution(tableName, varName)
}
\arguments{
\item{tableName}{Table name.}

\item{varName}{Variable name.}

\item{tableName}{table name.}

\item{varName}{variable name.}
}
\description{
Returns a single-row dataframe from catalog (udfCatalog) containing the
variable's spatial and temporal resolutions.

Returns a single-row dataframe from catalog (udfCatalog) containing the
variable's spatial and temporal resolutions.
}
\examples{
\dontrun{
tableName <- "tblArgoMerge_REP" # table name
varName <- "argo_merge_chl_adj" # variable name
get_var_resolution(tableName, varName)
}
\donttest{
library(cmap4r)
#
## Input: Table name; variable name
table_name <- "tblArgoMerge_REP" # table name
var_name <- "argo_merge_chl_adj" # variable name
#
## Variable attribute:
varResolution <- get_var_resolution(table_name, var_name)
varResolution
#
}
}
