% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_info.R
\name{get_var_stat}
\alias{get_var_stat}
\title{Returns a single-row dataframe from the database catalog containing the variable's summary statistics.}
\usage{
get_var_stat(tableName, varName)
}
\arguments{
\item{tableName}{table name from the Simons CMAP database. Use "get_catalog()" to retrieve list of tables on the database.}

\item{varName}{specify short name of a variable in the table.}
}
\value{
numerical attribute of a table variable as dataframe
}
\description{
Returns a single-row dataframe from the database catalog containing the variable's summary statistics.
}
\examples{
\donttest{
#
## Input: Table name; variable name
tableName <- "tblArgoMerge_REP" # table name
varName <- "argo_merge_chl_adj" # variable name
#
## Variable attribute:
varStats <- get_var_stat(tableName, varName)
varStats
#
}
}
