% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{has_field}
\alias{has_field}
\title{Returns a boolean confirming whether a field (varName) exists in a table (data set).}
\usage{
has_field(tableName, varName)
}
\arguments{
\item{tableName}{Table name.}

\item{varName}{Variable name.}
}
\description{
Returns a boolean confirming whether a field (varName) exists in a table (data set).
}
\examples{
\dontrun{
library(cmap4r)
#
## Input: Table name; variable name
table_name <- "tblArgoMerge_REP" # table name
var_name <- "argo_merge_chl_adj" # variable name
#
## Variable attribute:
var_exist <- has_field(table_name, var_name)
var_exist
#
}
}
