% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_info.R
\name{is_grid}
\alias{is_grid}
\title{Returns a boolean indicating whether the variable is a gridded product or has irregular spatial resolution.}
\usage{
is_grid(tableName, varName)
}
\arguments{
\item{tableName}{table name from the Simons CMAP database. Use "get_catalog()" to retrieve list of tables on the database.}

\item{varName}{specify short name of a variable in the table.}
}
\value{
boolean
}
\description{
Returns a boolean indicating whether the variable is a gridded product or has irregular spatial resolution.
}
\examples{
\donttest{

#
## Input:
table <- c("tblArgoMerge_REP") # table name
variable <- c("argo_merge_chl_adj") # variable name
#
is_grid(table, variable)

#
}
}
