% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_info.R
\name{is_grid}
\alias{is_grid}
\title{Returns a boolean indicating whether the variable is a gridded product or has irregular spatial resolution.}
\usage{
is_grid(table_name, var_name)
}
\arguments{
\item{table_name}{table name.}

\item{var_name}{variable name.}
}
\description{
Returns a boolean indicating whether the variable is a gridded product or has irregular spatial resolution.
}
\examples{
\donttest{
library(cmap4r)
#
## Input:
table <- c("tblArgoMerge_REP") # table name
variable <- c("argo_merge_chl_adj") # variable name
#
is_grid(table, variable)

#
}
}
