% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cmap.R
\name{plot_depth}
\alias{plot_depth}
\title{Plot variable of interest with varying depth}
\usage{
plot_depth(tbl_subset, type = c("plotly", "ggplot")[1], sel_var)
}
\arguments{
\item{tbl_subset}{table with depth and variable of interest}

\item{type}{choose a type of plot object: 'plotly','ggplot}

\item{sel_var}{select a variable in the table}
}
\value{
plot object of the chosen type
}
\description{
Depth plot object
}
\examples{
\dontrun{
library(cmap4r)
con <- connect_cmap(Driver = "libtdsodbc.so")
#
# Inpit variable:
table_list <- c("tblArgoMerge_REP", "tblPisces_NRT", "tblDarwin_Chl_Climatology")
var_list <- c("argo_merge_chl_adj", "CHL", "chl01_darwin_clim")
#
# selected argo_merge_chl_adj from tblArgoMerge_REP
selIndex <- 1
table_name <- table_list[selIndex]
sel_var <- var_list[selIndex]
#
range_var <- list()
range_var$lat <- c(20, 24)
range_var$lon <- c(-170, -150)
range_var$depth <- c(0, 1500)
range_var$time <- c("2016-04-30", "2016-04-30")
#
#
# Subset selection: data retrieval
agg_var <- "depth"
tbl_subset <- get_aggtable(con, table_name, sel_var, range_var, agg_var)
head(tbl_subset)
#
#
# Plot -- Depth profiles:
p <- plot_depth(tbl_subset, "plotly", sel_var)
#
#
dbDisconnect(con)
}
}
