% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cmap.R
\name{plot_depth}
\alias{plot_depth}
\title{Plot variable of interest with varying depth}
\usage{
plot_depth(tbl.subset, type = c("plotly", "ggplot")[1], sel.var)
}
\arguments{
\item{tbl.subset}{table with depth and variable of interest}

\item{type}{choose a type of plot object: 'plotly','ggplot}

\item{sel.var}{select a variable in the table}
}
\value{
plot object of the chosen type
}
\description{
Depth plot object
}
\examples{
\dontrun{
con <- connect2cmap(Driver = "libtdsodbc.so")
#
# Inpit variable:
table.list <- c("tblArgoMerge_REP", "tblPisces_NRT", "tblDarwin_Chl_Climatology")
var.list <- c("argo_merge_chl_adj", "CHL", "chl01_darwin_clim")
#
# selected argo_merge_chl_adj from tblArgoMerge_REP
selIndex <- 1
table.name <- table.list[selIndex]
sel.var <- var.list[selIndex]
#
range.var <- list()
range.var$lat <- c(20, 24)
range.var$lon <- c(-170, -150)
range.var$depth <- c(0, 1500)
range.var$time <- c("2016-04-30", "2016-04-30")
#
#
# Subset selection: data retrieval
agg.var <- "depth"
tbl.subset <- getAggregatedTableData(con, table.name, sel.var, range.var, agg.var)
head(tbl.subset)
#
#
# Plot -- Depth profiles:
p <- plot_depth(tbl.subset, "plotly", sel.var)
#
#
dbDisconnect(con)
}
}
