% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cmap.R
\name{plot_depth}
\alias{plot_depth}
\title{Plot variable of interest with varying depth}
\usage{
plot_depth(tableName, varName, lat1, lat2, lon1, lon2, dt1, dt2, depth1,
  depth2, type = c("plotly", "ggplot")[1], export_data = FALSE)
}
\arguments{
\item{tableName}{name of the table from CMAP}

\item{varName}{select a variable in the table}

\item{lat1}{start latitude [degree N].}

\item{lat2}{end latitude [degree N].}

\item{lon1}{start longitude [degree E].}

\item{lon2}{end longitude [degree E].}

\item{dt1}{start date or datetime.}

\item{dt2}{end date or datetime.}

\item{depth1}{start depth [m].}

\item{depth2}{end depth [m].}

\item{type}{choose a type of plot object: 'plotly','ggplot}

\item{export_data}{boolean variable to export data or not. The data will be saved in the working directory.}
}
\value{
plot object of the chosen type
}
\description{
Depth plot object
}
\examples{
\donttest{

#
# Inpit variable:
table_list <- c('tblArgoMerge_REP', 'tblPisces_NRT', 'tblDarwin_Ecosystem')
var_list <- c('argo_merge_chl_adj', 'CHL', 'CHL')
#
# selected argo_merge_chl_adj from tblArgoMerge_REP
selIndex <- 1
tableName <- table_list[selIndex]
varName <- var_list[selIndex]
#
# Range variable [lat,lon,time,depth]
lat1 = 20; lat2 = 24
lon1 = -170; lon2 = -150
dt1 = "2014-04-25"; dt2 = "2014-04-30"
depth1 <- 0; depth2 =  1500

#
# Plot -- Depth profiles:
p <- plot_depth(tableName, varName, lat1, lat2, lon1, lon2,
                        dt1, dt2,depth1, depth2, "plotly")
# p
#
}
}
