% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cmap.R
\name{plot_depth}
\alias{plot_depth}
\title{Plot variable of interest with varying depth}
\usage{
plot_depth(table_name, sel_var, range_var, type = c("plotly",
  "ggplot")[1], export_data = FALSE)
}
\arguments{
\item{table_name}{name of the table from CMAP}

\item{sel_var}{select a variable in the table}

\item{range_var}{range variable}

\item{type}{choose a type of plot object: 'plotly','ggplot}

\item{export_data}{boolean variable to export data or not. The data will be saved in the working directory.}
}
\value{
plot object of the chosen type
}
\description{
Depth plot object
}
\examples{
\donttest{
library(cmap4r)
#
# Inpit variable:
table_list <- c('tblArgoMerge_REP', 'tblPisces_NRT', 'tblDarwin_Ecosystem')
var_list <- c('argo_merge_chl_adj', 'CHL', 'CHL')
#
# selected argo_merge_chl_adj from tblArgoMerge_REP
selIndex <- 1
table_name <- table_list[selIndex]
sel_var <- var_list[selIndex]
#
range_var <- list()
range_var$lat <- c(20, 24)
range_var$lon <- c(-170, -150)
range_var$depth <- c(0, 1500)
range_var$time <- c("2014-04-25", "2014-04-30")
#
# Plot -- Depth profiles:
p <- plot_depth(table_name, sel_var, range_var, "plotly")
# p
#
}
}
