% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cmap.R
\name{plot_hist}
\alias{plot_hist}
\title{Plot histogram of the variable of interest}
\usage{
plot_hist(tbl_subset, type = c("plotly", "ggplot")[1], sel_var)
}
\arguments{
\item{tbl_subset}{table with depth and variable of interest}

\item{type}{choose a type of plot object: 'plotly','ggplot}

\item{sel_var}{select a variable in the table}
}
\value{
plot object of the chosen type
}
\description{
Histogram object of chosen type
}
\examples{
\dontrun{
library(cmap4r)
con <- connect_cmap(Driver = "libtdsodbc.so")
#
# Inpit variable:
table_list <- c("tblSST_AVHRR_OI_NRT", "tblArgoMerge_REP", "tblArgoMerge_REP")
var_list <- c("sst", "argo_merge_temperature_adj", "argo_merge_salinity_adj")
#
# variable "sst" selected from "tblSST_AVHRR_OI_NRT"
selIndex <- 1
table_name <- table_list[selIndex]
sel_var <- var_list[selIndex]
#
#
# Range variable
range_var <- list()
range_var$lat <- c(20, 24)
range_var$lon <- c(-170, 150)
range_var$time <- c("2016-04-30", "2016-04-30")
#
#
#
# Subset selection:
tbl_subset <- get_table(con, table_name, sel_var, range_var)
head(tbl_subset)
#
#
# Plot function available for R User
p <- plot_hist(tbl_subset, "ggplot", sel_var)
p <- plot_hist(tbl_subset, "plotly", sel_var)
p
#
#
dbDisconnect(con)
}
}
