% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cmap.R
\name{plot_hist}
\alias{plot_hist}
\title{Plot histogram of the variable of interest}
\usage{
plot_hist(tableName, varName, lat1, lat2, lon1, lon2, dt1, dt2,
  depth1 = NULL, depth2 = NULL, type = c("plotly", "ggplot")[1],
  export_data = FALSE)
}
\arguments{
\item{tableName}{name of the table from CMAP}

\item{varName}{select a variable in the table}

\item{lat1}{start latitude [degree N].}

\item{lat2}{end latitude [degree N].}

\item{lon1}{start longitude [degree E].}

\item{lon2}{end longitude [degree E].}

\item{dt1}{start date or datetime.}

\item{dt2}{end date or datetime.}

\item{depth1}{start depth [m].}

\item{depth2}{end depth [m].}

\item{type}{choose a type of plot object: 'plotly','ggplot}

\item{export_data}{boolean variable to export data or not. The data will be saved in the working directory.}
}
\value{
plot object of the chosen type
}
\description{
Histogram object of chosen type
}
\examples{
\donttest{

#
# Inpit variable:
table_list <- c("tblSST_AVHRR_OI_NRT", "tblArgoMerge_REP", "tblArgoMerge_REP")
var_list <- c("sst", "argo_merge_temperature_adj", "argo_merge_salinity_adj")
#
# variable "sst" selected from "tblSST_AVHRR_OI_NRT"
selIndex <- 1
tableName <- table_list[selIndex]
varName <- var_list[selIndex]
#
# Range variable [lat,lon,time,depth]
lat1 = 20; lat2 = 24
lon1 = -170; lon2 = -150
dt1 = "2016-04-30"; dt2 = "2016-04-30"
depth1 <- NULL; depth2 =  NULL


# Plot function available for R User
# p <- plot_hist(tableName, varName, lat1, lat2,
#                lon1, lon2, dt1, dt2, depth1, depth2, "ggplot")

p <- plot_hist(tableName, varName, lat1, lat2, lon1, lon2,
               dt1, dt2, depth1, depth2, "plotly")
p
#
}
}
