% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cmap.R
\name{plot_hist}
\alias{plot_hist}
\title{Histogram plot}
\usage{
plot_hist(tableName, varName, lat1, lat2, lon1, lon2, dt1, dt2,
  depth1 = NULL, depth2 = NULL, type = c("plotly", "ggplot")[1],
  export_data = FALSE)
}
\arguments{
\item{tableName}{table name from the Simons CMAP database. Use "get_catalog()" to retrieve list of tables on the database.}

\item{varName}{specify short name of a variable in the table. Use "get_catalog()" to retrieve list of table variables on the database.}

\item{lat1}{start latitude [degree N] of the meridional cut; ranges from -90° to 90°.}

\item{lat2}{end latitude [degree N] of the meridional cut; ranges from -90° to 90°.}

\item{lon1}{start longitude [degree E]  of the zonal cut; ranges from  -180° to 180°.}

\item{lon2}{end longitude [degree E] of the zonal cut; ranges from  -180° to 180°.}

\item{dt1}{start date or datetime (lower bound of temporal cut). Example values: '2016-05-25' or '2017-12-10 17:25:00'}

\item{dt2}{end date or datetime (upper bound of temporal cut). Example values: '2016-04-30' or '2016-04-30 17:25:00'}

\item{depth1}{positive value specifying the start depth [m] of the vertical cut. Note that depth  is 0 at surface and grows towards ocean floor.}

\item{depth2}{positive value specifying the end depth [m]of the vertical cut. Note that depth  is 0 at surface and grows towards ocean floor.}

\item{type}{choose a type of plot object: 'plotly','ggplot'. Default 'plotly'.}

\item{export_data}{boolean variable to export data or not. The data will be saved in the working directory. Default FALSE.}
}
\value{
plot object of the chosen type, i.e., plotly/ggplot
}
\description{
Plot histogram of the variable of interest after extracting the data (using get_spacetime) from the Simons CMAP database using the specified space-time constraints (dt1, dt2, lat1, lat2, lon1, lon2, depth1, depth2).
}
\examples{
\donttest{

#
# Inpit variable:
tableList <- c("tblSST_AVHRR_OI_NRT", "tblArgoMerge_REP", "tblArgoMerge_REP")
varList <- c("sst", "argo_merge_temperature_adj", "argo_merge_salinity_adj")
#
# variable "sst" selected from "tblSST_AVHRR_OI_NRT"
selIndex <- 1
tableName <- tableList[selIndex]
varName <- varList[selIndex]
#
# Range variable [lat,lon,time,depth]
lat1 = 20; lat2 = 24
lon1 = -170; lon2 = -150
dt1 = "2016-04-30"; dt2 = "2016-04-30"
depth1 <- NULL; depth2 =  NULL


# Plot function available for R User
# p <- plot_hist(tableName, varName, lat1, lat2,
#                lon1, lon2, dt1, dt2, depth1, depth2, "ggplot")

p <- plot_hist(tableName, varName, lat1, lat2, lon1, lon2,
               dt1, dt2, depth1, depth2, "plotly")
p
#
}
}
