% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cmap.R
\name{plot_hist}
\alias{plot_hist}
\title{Plot histogram of the variable of interest}
\usage{
plot_hist(tbl.subset, type = c("plotly", "ggplot")[1], sel.var)
}
\arguments{
\item{tbl.subset}{table with depth and variable of interest}

\item{type}{choose a type of plot object: 'plotly','ggplot}

\item{sel.var}{select a variable in the table}
}
\value{
plot object of the chosen type
}
\description{
Histogram object of chosen type
}
\examples{
\dontrun{
con <- dbConnect(odbc::odbc(), DSN="CMAP-MSSQL",UID="ArmLab",PWD="ArmLab2018")
#
# Inpit variable:
table.list <- c('tblSST_AVHRR_OI_NRT', 'tblArgoMerge_REP', 'tblArgoMerge_REP')
var.list <-  c('sst', 'argo_merge_temperature_adj', 'argo_merge_salinity_adj')
#
# variable "sst" selected from "tblSST_AVHRR_OI_NRT"
selIndex <- 1
table.name <- table.list[selIndex]
sel.var <- var.list[selIndex]
#
#
 # Range variable
range.var <- list()
range.var$lat <- c(20, 24)
range.var$lon <- c(-170, 150)
range.var$time <- c('2016-04-30', '2016-04-30')
#
#
#
# Subset selection:
tbl.subset <- getTableData(con, table.name, sel.var, range.var)
head(tbl.subset)
#
#
# Plot function available for R User
p <- plot_hist(tbl.subset,'ggplot',sel.var)
p <- plot_hist(tbl.subset,'plotly',sel.var)
p
#
#
dbDisconnect(con)
}
}
