% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cmap.R
\name{plot_regmap}
\alias{plot_regmap}
\title{Regional map}
\usage{
plot_regmap(con, table_name, sel_var, range_var, type = c("plotly",
  "ggplot")[1])
}
\arguments{
\item{con}{connection object to the database}

\item{table_name}{name of table from CMAP}

\item{sel_var}{selected varable from the table}

\item{range_var}{range of time, latitude, longitude, depth}

\item{type}{choose a type of plot object: 'plotly','ggplot}
}
\value{
\item{plot}{ plot object of the chosen type}
   \item{table}{subset of data from  first table }
}
\description{
Regional map of a variable of interest at varying range of latitude and longitude
}
\examples{
\dontrun{
library(cmap4r)
con <- connect_cmap(Driver = "libtdsodbc.so")
#
# Inpit variable:
table_name <- "tblsst_AVHRR_OI_NRT"
sel_var <- "sst"
range_var <- list()
range_var$lat <- c(10, 70)
range_var$lon <- c(-180, -80)
range_var$time <- c("2016-04-30", "2016-04-30")
#
## Plot - regional map
# out <- plot_regmap(con, table_name,sel_var,range_var, type = 'ggplot')
out <- plot_regmap(con, table_name, sel_var, range_var, type = "plotly")
out$plot
}
}
