% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cmap.R
\name{plot_regMap}
\alias{plot_regMap}
\title{Regional map}
\usage{
plot_regMap(con, table.name, sel.var, range.var, type = c("plotly",
  "ggplot")[1])
}
\arguments{
\item{con}{connection object to the database}

\item{table.name}{name of table from CMAP}

\item{sel.var}{selected varable from the table}

\item{range.var}{range of time, latitude, longitude, depth}

\item{type}{choose a type of plot object: 'plotly','ggplot}
}
\value{
\item{plot}{ plot object of the chosen type}
   \item{table}{subset of data from  first table }
}
\description{
Regional map of a variable of interest at varying range of latitude and longitude
}
\examples{
\dontrun{
con <- dbConnect(odbc::odbc(), DSN="CMAP-MSSQL",UID="ArmLab",PWD="ArmLab2018")
#
# Inpit variable:
table.name = 'tblsst_AVHRR_OI_NRT'
sel.var = 'sst'
range.var <- list()
range.var$lat <- c(10,70)
range.var$lon <- c(-180,-80)
range.var$time <- c('2016-04-30', '2016-04-30')
#
## Plot - regional map
# out <- plot_regMap(con, table.name,sel.var,range.var, type = 'ggplot')
out <- plot_regMap(con, table.name,sel.var,range.var, type = 'plotly')
out$plot
}
}
