% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cmap.R
\name{plot_regmap}
\alias{plot_regmap}
\title{Regional map}
\usage{
plot_regmap(tableName, varName, lat1, lat2, lon1, lon2, dt1, dt2,
  depth1 = NULL, depth2 = NULL, type = c("plotly", "ggplot")[1],
  export_data = FALSE)
}
\arguments{
\item{tableName}{name of table from CMAP}

\item{varName}{selected varable from the table}

\item{lat1}{start latitude [degree N].}

\item{lat2}{end latitude [degree N].}

\item{lon1}{start longitude [degree E].}

\item{lon2}{end longitude [degree E].}

\item{dt1}{start date or datetime.}

\item{dt2}{end date or datetime.}

\item{depth1}{start depth [m].}

\item{depth2}{end depth [m].}

\item{type}{choose a type of plot object: 'plotly','ggplot}

\item{export_data}{boolean variable to export data or not. The data will be saved in the working directory.}
}
\value{
plot object of the chosen type
}
\description{
Regional map of a variable of interest at varying range of latitude and longitude
}
\examples{
\donttest{

#
# Inpit variable:
tableName <- "tblsst_AVHRR_OI_NRT"
varName <- "sst"

# Range variable [lat,lon,time,depth]
lat1 = 10; lat2 = 70
lon1 = -180; lon2 = -80
dt1 = "2016-04-30"; dt2 = "2016-04-30"
depth1 <- NULL; depth2 =  NULL

#
## Plot - regional map
# out <- plot_regmap( tableName,varName, lat1, lat2, lon1, lon2,
#                    dt1, dt2, depth1, depth2, type = 'ggplot')
out <- plot_regmap(tableName, varName, lat1, lat2, lon1, lon2,
                   dt1, dt2, depth1, depth2, type = "plotly")
out
}
}
