% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cmap.R
\name{plot_ts}
\alias{plot_ts}
\title{Plot variable of interest with varying time.}
\usage{
plot_ts(table_name, sel_var, lat1, lat2, lon1, lon2, dt1, dt2,
  depth1 = NULL, depth2 = NULL, type = c("plotly", "ggplot")[1],
  export_data = FALSE)
}
\arguments{
\item{table_name}{table with depth and variable of interest}

\item{sel_var}{select a variable in the table}

\item{lat1}{start latitude [degree N].}

\item{lat2}{end latitude [degree N].}

\item{lon1}{start longitude [degree E].}

\item{lon2}{end longitude [degree E].}

\item{dt1}{start date or datetime.}

\item{dt2}{end date or datetime.}

\item{depth1}{start depth [m].}

\item{depth2}{end depth [m].}

\item{type}{choose a type of plot object: 'plotly','ggplot}

\item{export_data}{boolean variable to export data or not. The data will be saved in the working directory.}
}
\value{
plot object of the chosen type
}
\description{
Times eries plot object.
}
\examples{
\donttest{
library(cmap4r)
#
# Input variable:
table_list <- c("tblSST_AVHRR_OI_NRT", "tblAltimetry_REP", "tblPisces_NRT")
var_list <- c("sst", "sla", "NO3")
#
# selected "sst" from the table "tblSST_AVHRR_OI_NRT"
selIndex <- 1
table_name <- table_list[selIndex]
sel_var <- var_list[selIndex]
#
# # Example I:
# Range variable [lat,lon,time,depth]
lat1 = 25; lat2 = 30
lon1 = -160; lon2 = -155
dt1 = "2016-03-29"; dt2 = "2016-05-29"
#
# Plot -- Time series:
p <- plot_ts(table_name, sel_var, lat1, lat2, lon1, lon2,
             dt1, dt2)
p
#
}
}
