% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cmap.R
\name{plot_ts}
\alias{plot_ts}
\title{Plot variable of interest with varying time}
\usage{
plot_ts(tbl.subset, type = c("plotly", "ggplot")[1], sel.var)
}
\arguments{
\item{tbl.subset}{table with depth and variable of interest}

\item{type}{choose a type of plot object: 'plotly','ggplot}

\item{sel.var}{select a variable in the table}
}
\value{
plot object of the chosen type
}
\description{
times eries plot object
}
\examples{
\dontrun{
con <- dbConnect(odbc::odbc(), DSN="CMAP-MSSQL",UID="ArmLab",PWD="ArmLab2018")
#
# Input variable:
table.list <- c('tblSST_AVHRR_OI_NRT', 'tblAltimetry_REP', 'tblPisces_NRT')
var.list <- c('sst', 'sla', 'NO3')
#
# selected "sst" from the table "tblSST_AVHRR_OI_NRT"
selIndex <- 1
table.name <- table.list[selIndex]
sel.var <- var.list[selIndex]
#
# # Example I:
range.var <- list()
range.var$lat <- c(25,30)
range.var$lon <- c(-160,-155)
range.var$time <- c('2016-03-29', '2016-05-29')
#
 # Subset selection: data retrieval
agg.var <- 'time'
tbl.subset <- getAggregatedTableData(con, table.name, sel.var, range.var, agg.var)
head(tbl.subset,20)
#
# Plot -- Time series:
p <- plot_ts(tbl.subset,'plotly',sel.var)
p
#
dbDisconnect(con)
}
}
