% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cmap.R
\name{plot_xy}
\alias{plot_xy}
\title{XY plot}
\usage{
plot_xy(table_list, var_list, range_var, agg_var, type = c("plotly",
  "ggplot")[1], export_data = FALSE)
}
\arguments{
\item{table_list}{list of  tables}

\item{var_list}{list of  corresponding variable}

\item{range_var}{range of time, latitude, longitude, depth}

\item{agg_var}{aggregate variable}

\item{type}{choose a type of plot object: 'plotly','ggplot}

\item{export_data}{boolean variable to export data or not. The data will be saved in the working directory.}
}
\value{
plot object of the chosen type
}
\description{
Scatter plot of two variable of interest
}
\examples{
\donttest{
library(cmap4r)
# Inpit variable:
#
table_list <- c("tblSST_AVHRR_OI_NRT", "tblAltimetry_REP")
var_list <- c("sst", "sla")
#
range_var <- list()
range_var$lat <- c(25, 30)
range_var$lon <- c(-160, -155)
range_var$time <- c("2016-03-29", "2016-05-29")
#
agg_var <- "time"

# xy plot
out <- plot_xy(table_list, var_list, range_var, agg_var)
out
}
}
